swaped vector from int to float [stable]

+++ includes/structures/cube_vector.h
@@ -15,16 +15,16 @@
 
 typedef struct s_vector
 {
-    int x;
-    int y;
+    float x;
+    float y;
 }               t_vector;
 
 /*
 **  CREATORS & DESTRUCTORS
 */
 
-t_vector    create_vector(int x, int y);
-t_vector    *malloc_vector(int x, int y);
+t_vector    create_vector(float x, float y);
+t_vector    *malloc_vector(float x, float y);
 void        free_vector(t_vector *to_free);
 
 #endif
