+++ srcs/structures/vector/creators_destructors/create_vector.c
@@ -12,7 +12,7 @@
 
 #include "cube.h"
 
-t_vector    create_vector(int x, int y)
+t_vector    create_vector(float x, float y)
 {
     t_vector    result;
