+++ srcs/structures/vector/creators_destructors/malloc_vector.c
@@ -12,7 +12,7 @@
 
 #include "cube.h"
 
-t_vector    *malloc_vector(int x, int y)
+t_vector    *malloc_vector(float x, float y)
 {
     t_vector    *result;
